pyproject.toml
[project]
name = "graphrag-poc"
version = "0.1.0"
description = "GraphRAG Chatbot PoC - Knowledge graph + vector search hybrid RAG system"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "qdrant-client>=1.12.0",
    "nebula3-python>=3.4.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.12",
    "pypdf>=6.10.1",
    "streamlit>=1.38.0",
    "streamlit-agraph>=0.0.5",
    "httpx>=0.27.0",
    "google-adk[extensions]>=1.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[project.scripts]
serve = "app.main:run_server"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
